requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic-settings==2.3.4
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==23.2.1
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0

# LangChain
langchain>=0.2.0,<0.3
langchain-ollama>=0.1.0    # ChatOllama + OllamaEmbeddings (replaces langchain-openai)
langchain-community>=0.2.0,<0.3
langchain-core>=0.2.0,<0.3

# Vector & Retrieval
faiss-cpu>=1.9.0
rank-bm25==0.2.2
numpy==1.26.4

# Document Loaders
pymupdf==1.24.5            # PyMuPDFLoader
unstructured==0.14.6       # HTML / other formats

# Smart Chunking & Graph
nltk==3.8.1
networkx==3.3
sentence-transformers==3.0.1

# Sandbox
RestrictedPython==7.1

# HTTP
httpx==0.27.0
requests>=2.28.0

# Redis (optional)
redis>=5.0.0

# Dev / CI
pytest==8.2.2
pytest-asyncio==0.23.7
ruff==0.4.10
mypy==1.10.0
